add middleware for admin

+++ routes/web.php
@@ -25,8 +25,8 @@ Route::post('logout',[SessionsController::class,'destroy'])->middleware('auth');
 Route::get('login',[SessionsController::class,'create'])->middleware('guest');
 Route::post('sessions',[SessionsController::class,'store'])->middleware('guest');
 
-Route::get('admin/posts/create',[PostController::class,'create']);
-Route::post('admin/posts',[PostController::class,'store']);
-Route::get('admin/posts/{post}/edit',[PostController::class,'edit']);
-Route::patch('admin/posts/{post}',[PostController::class,'update']);
-Route::delete('admin/posts/{post}',[PostController::class,'destroy']);
+Route::get('admin/posts/create',[PostController::class,'create'])->middleware('can:admin');
+Route::post('admin/posts',[PostController::class,'store'])->middleware('can:admin');
+Route::get('admin/posts/{post}/edit',[PostController::class,'edit'])->middleware('can:admin');
+Route::patch('admin/posts/{post}',[PostController::class,'update'])->middleware('can:admin');
+Route::delete('admin/posts/{post}',[PostController::class,'destroy'])->middleware('can:admin');
